Move interval runner package from redisreceiver to internal/interval
