+++ internal/interval/go.mod
@@ -1,0 +1,3 @@
+module github.com/open-telemetry/opentelemetry-collector-contrib/interval
+
+go 1.16
