+++ receiver/dockerstatsreceiver/go.mod
@@ -8,13 +8,14 @@ require (
 	github.com/gobwas/glob v0.2.3
 	github.com/golang/protobuf v1.5.2
 	github.com/open-telemetry/opentelemetry-collector-contrib/internal/common v0.0.0-00010101000000-000000000000
-	github.com/open-telemetry/opentelemetry-collector-contrib/receiver/redisreceiver v0.0.0-00010101000000-000000000000
+	github.com/open-telemetry/opentelemetry-collector-contrib/internal/interval v0.0.0-00010101000000-000000000000
 	github.com/stretchr/testify v1.7.0
 	go.opentelemetry.io/collector v0.31.1-0.20210810171211-8038673eba9e
 	go.uber.org/zap v1.19.0
 	google.golang.org/protobuf v1.27.1
 )
 
-replace github.com/open-telemetry/opentelemetry-collector-contrib/internal/common => ../../internal/common
-
-replace github.com/open-telemetry/opentelemetry-collector-contrib/receiver/redisreceiver => ../redisreceiver
+replace (
+	github.com/open-telemetry/opentelemetry-collector-contrib/internal/common => ../../internal/common
+	github.com/open-telemetry/opentelemetry-collector-contrib/internal/interval => ../../internal/interval
+)
