+++ receiver/redisreceiver/go.mod
@@ -7,10 +7,14 @@ require (
 	github.com/onsi/ginkgo v1.14.1 // indirect
 	github.com/onsi/gomega v1.10.2 // indirect
 	github.com/open-telemetry/opentelemetry-collector-contrib/internal/common v0.0.0-00010101000000-000000000000
+	github.com/open-telemetry/opentelemetry-collector-contrib/internal/interval v0.0.0-00010101000000-000000000000
 	github.com/stretchr/testify v1.7.0
 	go.opentelemetry.io/collector v0.31.1-0.20210810171211-8038673eba9e
 	go.opentelemetry.io/collector/model v0.31.1-0.20210810171211-8038673eba9e
 	go.uber.org/zap v1.19.0
 )
 
-replace github.com/open-telemetry/opentelemetry-collector-contrib/internal/common => ../../internal/common
+replace (
+	github.com/open-telemetry/opentelemetry-collector-contrib/internal/common => ../../internal/common
+	github.com/open-telemetry/opentelemetry-collector-contrib/internal/interval => ../../internal/interval
+)
