+++ receiver/redisreceiver/redis_runnable.go
@@ -24,7 +24,7 @@ import (
 	"go.opentelemetry.io/collector/obsreport"
 	"go.uber.org/zap"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/redisreceiver/interval"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/interval"
 )
 
 var _ interval.Runnable = (*redisRunnable)(nil)
